pyproject.toml
[project]
name = "davidkhala.utils"
version = "0.6.0"
description = "@davidkhala/python-utils"
authors = [{ name = "David Liu", email = "david-khala@hotmail.com" }]
requires-python = ">=3.11"
readme = "README.md"

[project.optional-dependencies]
sql = ["sqlparse"]
http_request = ["requests"]
poetry = ["python-dotenv"]
build = ["pyinstaller ; python_version < '3.14'"]

[project.scripts]
build-poetry = "poetry:build"
clean-poetry = "poetry:clean"

[dependency-groups]
dev = [
    "pytest",
]

[tool.hatch.build.targets.sdist]
include = ["davidkhala"]

[tool.hatch.build.targets.wheel]
include = ["davidkhala"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
